requirements.txt
streamlit
sympy
matplotlib
numpy
python-dotenv
openai

